Handle edge cases for $ environment variable
