+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: ogonzale <ogonzale@student.42barcel>       +#+  +:+       +#+         #
+#    By: oscar <oscar@student.42.fr>                +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/05/18 10:00:13 by ogonzale          #+#    #+#              #
-#    Updated: 2022/12/03 18:01:37 by ogonzale         ###   ########.fr        #
+#    Updated: 2022/12/06 12:52:29 by oscar            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -40,7 +40,7 @@ WHITE 		:= \033[0;97m
 
 PARSE_DIR	:= parse/
 PARSE_FILES	:= split_cmd_line ft_split_mod split_words split_words_2 split_words_3\
-	           split_utils split_utils_2 expand_words expand_words_2
+	           split_utils split_utils_2 expand_words expand_words_2 expand_words_3
 
 MAND_DIR	:= mand/
 MAND_FILES	+= $(addprefix $(PARSE_DIR), $(PARSE_FILES))
