+++ inc/minishell.h
@@ -6,14 +6,22 @@
 /*   By: ogonzale <ogonzale@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 16:36:03 by ogonzale          #+#    #+#             */
-/*   Updated: 2022/12/03 18:01:19 by ogonzale         ###   ########.fr       */
+/*   Updated: 2022/12/07 11:11:59 by ogonzale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef MINISHELL_H
 # define MINISHELL_H
 
-# include "libft.h"
+#include "libft.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <unistd.h>
+#include <readline/readline.h>
+#include <readline/history.h>
+#include <sys/types.h>
+#include <sys/wait.h>
+#include <signal.h>
 
 typedef struct s_cmd_line_content
 {
@@ -87,6 +95,10 @@ int		expand_words(t_list **cmd_line);
 /* expand_words_2.c */
 
 void	expand(char *word, int *remove_char);
+
+/* expand_words_3.c */
+
+char	*expand_env(char *word);
 
 /* free.c */
